Revert "First attempt at polymorphic old expression"
This reverts commit 827c21f3406c1e151f2d812d99a0df470880e2bb.

+++ test/solc-verify/OldExpr.sol
@@ -2,26 +2,24 @@ pragma solidity >=0.5.0;
 
 contract OldExpr {
     address owner;
-    int8 counter;
+    int256 counter;
 
     constructor() public {
         owner = msg.sender;
     }
 
-    // Correct version, only owner can change the owner
-    /** @notice postcondition msg.sender == __verifier_old(owner) || owner == __verifier_old(owner) */
+    /** @notice postcondition msg.sender == __verifier_old_address(owner) || owner == __verifier_old_address(owner) */
     function changeOwnerCorrect(address newOwner) public {
         require(msg.sender == owner);
         owner = newOwner;
     }
 
-    // Incorrect version, anyone can change the owner
-    /** @notice postcondition msg.sender == __verifier_old(owner) || owner == __verifier_old(owner) */
+    /** @notice postcondition msg.sender == __verifier_old_address(owner) || owner == __verifier_old_address(owner) */
     function changeOwnerIncorrect(address newOwner) public {
         owner = newOwner;
     }
 
-    /** @notice postcondition counter == __verifier_old(counter) + 1 */
+    /** @notice postcondition counter == __verifier_old_int256(counter) + 1 */
     function increment() public {
         counter++;
     }
